refactor: remove XDSSideNavCollapseProvider, use context directly
React 19 supports <Context value={...}> without .Provider. Remove
the unnecessary Provider wrapper — use XDSSideNavCollapseContext
directly in SideNavItem popover and test helpers.

CollapseContext file stays as .ts (no JSX needed).
